BIG UPDATE: no new features, but code improved
The feature of ArtClass are the same as before: you give a source file in input (of one of the supported languages) and you get the same file but refactored in output. Some old modules were modified and others were deleted. The code readability is improved.
Multiline strings support is still a missing piece due to the very different implementations in the various languages.